requirements.txt
flask
ibm_db
ibm-cos-sdk
pillow